projects/MarketAura/pyproject.toml
[project]
name = "marketaura"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nadeem", email = "nadeemsangrasi903@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai-agents[litellm]>=0.0.19",
    "chainlit>=2.5.5",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langmem>=0.0.27",
    "langmem-adapter>=0.0.2",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0"
]

[project.scripts]
agent = "marketaura.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
